requirements.txt
aiohttp==3.7.4
async-lru==1.0.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.1
bidict==0.21.2
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
chucknorris==2.8
youtube_dl
py-kingbot
youtube-search-python
google==3.0.0
google-search==1.1.1
googlesearch-python==2020.0.2
greenlet==1.1.0
idna==2.10
joke-generator==1.0.1
lxml==4.6.3
multidict==5.1.0
Naked==0.1.31
nekobin==0.0.3
Pastebin==1.1.2
psutil==5.8.0
py-tgcalls==0.5.0
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
pyspeedtest==1.2.7
python-decouple==3.4
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
PyYAML==5.4.1
redis==3.5.3
requests==2.24.0
soupsieve==2.0.1
speedtest-cli==2.1.3
SQLAlchemy==1.4.15
telegraph==1.4.1
psycopg2
TgCrypto==1.2.2
typing-extensions==3.10.0.0
urllib3==1.25.9
yarl==1.6.3
gitpython>=3.1.14
heroku3>=4.2.3
